+++ app/Resources/doc/changelog.md
@@ -2,9 +2,14 @@
 
 ## 2.0.4.0
 * Helptext import/export commands.
+* Added Searchbundle.
+* Changed metadata permissions.
+
+## 2.0.3.3
+* Minor bugfixes.
 
 ## 2.0.3.1
-* Added position name to address wizard
+* Added position name to address wizard.
 * Refactored metadata template structure.
 
 ## 2.0.3.0
@@ -16,6 +21,5 @@
 
 ## 2.0.2.1
 * Upgrade Symfony version (2.3.7).
-
 
 <a href="../../../README.md">&laquo; back</a>